pyproject.toml
[tool.poetry]
name = "langgraph-agent"
version = "0.1.0"
description = ""
authors = ["Carlos L. <krlosaren@gmail.com> <krlosaren@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pymysql = "^1.1.1"
python-dotenv = "^1.0.1"
loguru = "^0.7.3"
langgraph = "^0.2.74"
langsmith = "^0.3.10"
langchain-openai = "^0.3.6"
rich = "^13.9.4"
tavily-python = "^0.5.1"
langchain-community = "^0.3.18"
langchain = "^0.3.19"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
